+++ src/main/java/io/hhplus/tdd/point/business/UserPointService.java
@@ -1,6 +1,6 @@
 package io.hhplus.tdd.point.business;
 
-import io.hhplus.tdd.point.dto.UserPointSelectDTO;
+import io.hhplus.tdd.point.business.dto.UserPointSelectDTO;
 
 public interface UserPointService {
 	UserPointSelectDTO findById(long id);
