+++ src/main/java/io/hhplus/tdd/point/business/UserPointServiceImpl.java
@@ -3,7 +3,7 @@ package io.hhplus.tdd.point.business;
 import org.springframework.stereotype.Service;
 
 import io.hhplus.tdd.common.api.support.error.ErrorType;
-import io.hhplus.tdd.point.dto.UserPointSelectDTO;
+import io.hhplus.tdd.point.business.dto.UserPointSelectDTO;
 import io.hhplus.tdd.point.exception.InvalidChargingPointException;
 import io.hhplus.tdd.point.exception.InvalidUserIdException;
 import io.hhplus.tdd.point.exception.UserNotFoundException;
