+++ src/main/java/io/hhplus/tdd/point/business/dto/UserPointSelectDTO.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point.dto;
+package io.hhplus.tdd.point.business.dto;
 
 public record UserPointSelectDTO(long userId, long point) {
 }
