+++ src/test/java/io/hhplus/tdd/point/presentation/PointControllerChargeUnitTest.java
@@ -17,7 +17,7 @@ import org.springframework.test.web.servlet.ResultActions;
 
 import io.hhplus.tdd.common.api.support.error.ErrorType;
 import io.hhplus.tdd.point.business.UserPointService;
-import io.hhplus.tdd.point.dto.UserPointSelectDTO;
+import io.hhplus.tdd.point.business.dto.UserPointSelectDTO;
 import io.hhplus.tdd.point.exception.InvalidChargingPointException;
 import io.hhplus.tdd.point.exception.InvalidUserIdException;
 import io.hhplus.tdd.point.exception.UserNotFoundException;
